Reverting the change in the previous commit.
As written in the comments of the revert commit in the PS2SDK repository:
I modified PS2Ident to show the entire 16-bit revision register, and it reports its revision as 0x00FF. Rolling back to the previous revision because that revision was working correctly.

It IS debatable whether it's worth keeping a workaround for this broken adaptor, but I've decided that there's no harm in providing a workaround for it. If someday, somebody releases a good SATA network adaptor, please feel free to entirely remove this workaround!